Border as a 'tool'
Set up a potential version of border as a left sidebar tool. This version creates a 12pt border using the current selected color. Changing the size while the border tool is selected will apply the size change to *future* uses of the border tool. Definitely not a polished final product.