IME input compatible
The issue is as follows:
The text input needs to be compatible with IME string input.

Here's a brief overview of the problem:
For languages like English, strings are directly input into the Input field. However, for some languages (such as Japanese and Chinese), input is done through an IME. The IME treats the input string as a buffer. Since this buffer passes the entire string being input each time the input event is called, the current process results in the string being displayed in duplicate.

For example, when typing "ABC", the Input fires events for "A", "AB", "ABC". As a result, "AABABC" is entered on the screen (this example uses English characters for illustration purposes). This can be critical for countries that standardly use an IME.

The correction is as follows:
When IME input begins, a compositionstart event is called, and when it ends, a compositionend event is called. For IME inputs, since the Input event stores the entire string in the buffer, I've concatenated the pre-held string before input and the string being input, replacing everything accordingly. There might be a better method, but I believe replacing everything should not be particularly heavy processing.